pyproject.toml
[tool.poetry]
name = "xshop-web"
version = "0.1.0"
description = "XShop's web-related functionality (Dashboard, API, etc...)"
authors = ["Ahmed Shahwan <ahmed@shahwan.me>"]

[tool.poetry.dependencies]
python = "3.8.6"
django = "3.1.*"
django-phonenumber-field = {extras = ["phonenumbers"], version = "4.0.0"}
model_bakery = "1.1.1"
django-model-utils = "4.0.0"
django-extensions = "3.0.5"
django-environ = "0.4.5"
psycopg2-binary = "2.8.5"
gunicorn = "20.0.4"
whitenoise = "5.2.0"
sentry-sdk = "0.17.3"
django-multiselectfield = "0.1.12"
django-jazzmin = "2.4.0"
django-money = "1.1"
djangorestframework = "3.12.1"
PyYAML = "5.3.1"
uritemplate = "3.0.1"
Markdown = "3.3"
Pygments = "2.7.2"
django-filter = "2.4.0"
django-crispy-forms = "1.9.2"
django-admin-list-filter-dropdown = "1.0.3"
drf-yasg2 = "1.19.3"
WeasyPrint = "51"
django-baton = "1.12.1"
pyOpenSSL = "19.1.0"
python-barcode = "0.13.1"
crispy-tailwind = "^0.2.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8-black = "0.2.1"
ipdb = "0.13.3"
docker-compose = "1.27.4"
flake9 = "3.8.3"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
    | manage.py,
    | venv,
    | \.venv,
    | \.env,
    | env,
    | downloads
    | static
    | templates
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.flake8]
max-line-length = 88
max-complexity = 25
select = 'C,E,F,W,B,B950'
ignore = 'E133,E501,W503,E203,W605'
hang-closing = true
exclude = '''
    .git,
    migrations,
    __pycache__,
    manage.py,
    templates,
    dist,
    build,
    venv,
    .venv,
    .env,
    env,
    node_modules,
    .svn,
    CVS,
    .bzr,
    .hg,
    .git,
    __pycache__,
    .tox,
    .eggs,
    *.egg
'''
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

